add project base

+++ student-packages/PROJECT-REGULATION/utils/include/utils.h
@@ -5,7 +5,7 @@
 #include "pico/time.h"
 
 // Macro to convert milliseconds to microseconds
-#define MS_TO_US(x) (uint32_t)((x) * 1e3)
+#define MS_TO_US(x) (uint64_t)((x) * 1e3)
 
 // Macro to convert microseconds to milliseconds
 #define US_TO_MS(x) (float_t)((x) / 1e3)
@@ -20,7 +20,7 @@
 #define US_TO_S(x) (float_t)((x) / 1e6)
 
 // Macro to convert seconds to microseconds
-#define S_TO_US(x) (uint32_t)((x) * 1e6)
+#define S_TO_US(x) (uint64_t)((x) * 1e6)
 
 // Macro to crop a value between a minimum and maximum
 #define CROP(x, min, max) (x < min ? min : (x > max ? max : x))
@@ -29,6 +29,6 @@
 uint32_t millis();
 
 // Function to get the current time in microseconds
-uint32_t micros();
+uint64_t micros();
 
 #endif // UTILS_H
